backend/requirements.txt
# Schwab API (requires Python 3.10+)
schwab-py>=1.5.1

# Supabase
supabase>=2.11.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.1

# HTTP Requests (for Polygon.io REST API)
requests>=2.31.0

# Type Checking
mypy>=1.5.0
types-requests

# Logging
structlog>=23.1.0

# Date/Time
python-dateutil>=2.8.2
pytz>=2023.3

